chore: added announcements CRUD
Announcements API and console view are now both working. This
commit changed up the file structure quite a bit to make it more
intuitive. Components are now stored in their own folder, API
handlers and types are stored in the lib folder, and courses and
announcements have their own folders as well. As for the logic,
the console page can switch between course view and announcements
view through the button on the top right.
