+++ lib/types/types.ts
@@ -1,0 +1,12 @@
+export type announcementType = {
+  id: string;
+  title: string;
+  link: string;
+  description: string;
+};
+
+export type courseType = {
+  id: string;
+  titleLong: string;
+  description: string;
+};
