+++ pages/api/announcements/[id].ts
@@ -1,0 +1,6 @@
+import all_handler from '../../../lib/handler/[id]';
+import { NextApiRequest, NextApiResponse } from 'next';
+
+export default async function handler(req: NextApiRequest, res: NextApiResponse) {
+  return all_handler(req, res, "announcements", "title");
+}
